♻️Refactor: 코드 컴포넌트화
