+++ src/Components/container/manager/VacationStatisticContainer.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const VacationStatisticContainer = ({ children }: { children: React.ReactNode }) => {
+  return <div className="flex flex-col gap-4">{children}</div>;
+};
+
+export default VacationStatisticContainer;
